+++ Genius/Models/Maintenance/MaintenanceCheck.swift
@@ -23,9 +23,11 @@ struct MaintenanceCheck<
 			Defaults[.developmentMode] ? Symbol(.minus, color: .primary, label: .notAvailable) : .none
 		} else if let syncValue {
 			if let value = syncValue.optional {
-				value >= requirement ? // swiftlint:disable:this void_function_in_ternary
-					Symbol(.checkmark, color: .green, label: .passed) :
+				if value >= requirement {
+					Defaults[.showPassedMaintenanceChecks] ? Symbol(.checkmark, color: .green, label: .passed) : nil
+				} else {
 					Symbol(.xmark, color: .red, label: .failed)
+				}
 			} else {
 				Defaults[.developmentMode] || Defaults[.interfaceMode] >= .advanced ?
 					Symbol(.questionmark, color: .red, label: .unknown) :
@@ -40,9 +42,11 @@ struct MaintenanceCheck<
 		if !?available ?? false {
 			Defaults[.developmentMode] ? Symbol(.minus, color: .primary, label: .notAvailable) : nil
 		} else if let value = await value.optional {
-			value >= requirement ? // swiftlint:disable:this void_function_in_ternary
-				Symbol(.checkmark, color: .green, label: .passed) :
+			if value >= requirement {
+				Defaults[.showPassedMaintenanceChecks] ? Symbol(.checkmark, color: .green, label: .passed) : nil
+			} else {
 				Symbol(.xmark, color: .red, label: .failed)
+			}
 		} else {
 			Defaults[.developmentMode] || Defaults[.interfaceMode] >= .advanced ?
 				Symbol(.questionmark, color: .red, label: .unknown) : nil