+++ GeniusTests/Extensions/Genius/MaintenanceCheck.swift
@@ -14,9 +14,11 @@ extension MaintenanceCheck where ValueWrapper == SyncValueWrapper<Value> {
 		if !?available ?? false {
 			Defaults[.developmentMode] ? Symbol(.minus, color: .primary, label: .notAvailable) : nil
 		} else if let value = value.optional {
-			value >= requirement ? // swiftlint:disable:this void_function_in_ternary
-				Symbol(.checkmark, color: .green, label: .passed) :
+			if value >= requirement {
+				Defaults[.showPassedMaintenanceChecks] ? Symbol(.checkmark, color: .green, label: .passed) : nil
+			} else {
 				Symbol(.xmark, color: .red, label: .failed)
+			}
 		} else {
 			Defaults[.developmentMode] || Defaults[.interfaceMode] >= .advanced ?
 				Symbol(.questionmark, color: .red, label: .unknown) : nil